requirements.txt
fastapi==0.78.0
aiopg==1.3.4
sqlalchemy==1.4.39
black
isort
autoflake
uvicorn~=0.23.2
pydantic~=1.10.12
starlette~=0.19.1
pyjwt==2.8.0